Fix prompt strategies
