+++ apps/interactor/src/libs/memory/strategies/AbtractPromptStrategy.ts
@@ -1,25 +1,235 @@
-import { MikuCard } from '@mikugg/bot-utils'
+import {
+  EMOTION_GROUP_TEMPLATES,
+  EMPTY_MIKU_CARD,
+  MikuCard,
+} from '@mikugg/bot-utils'
 import { RootState } from '../../../state/store'
-import { NarrationResponse } from '../../../state/versioning'
+import {
+  NarrationInteraction,
+  NarrationResponse,
+} from '../../../state/versioning'
 import llamaTokenizer from '../_llama-tokenizer'
+import { fillTextTemplate, parseLLMResponse } from '.'
+import {
+  selectCurrentScene,
+  selectCurrentCharacterOutfits,
+  selectLastLoadedCharacters,
+  selectAllParentDialogues,
+} from '../../../state/selectors'
+
+const EMOTION_TOKEN_OFFSET = 4
 
 export abstract class AbstractPromptStrategy {
-  public abstract buildPrompt(
+  protected abstract getContextPrompt(
+    state: RootState,
+    currentRole: string
+  ): string
+
+  protected abstract template(): {
+    instruction: string
+    response: string
+    askLine: string
+    stops: string[]
+  }
+
+  public buildPrompt(
     state: RootState,
     maxNewTokens: number,
     memorySize: number,
-    role: string
+    currentRole: string
   ): {
     template: string
     variables: Record<string, string | string[]>
     totalTokens: number
+  } {
+    const roles = selectCurrentScene(state)?.roles || []
+    const outfits = selectCurrentCharacterOutfits(state)
+    const charactedId =
+      outfits.find(({ role }) => role === currentRole)?.id || ''
+    const { name } = this.getCharacterSpecs(
+      state.novel.characters[charactedId]?.card || EMPTY_MIKU_CARD
+    )
+    const emotions = this.getRoleEmotions(state, currentRole)
+
+    let template = this.getContextPrompt(state, currentRole)
+    template += this.getDialogueHistoryPrompt(state, memorySize, currentRole)
+    template += this.getResponseAskLine(state, maxNewTokens, currentRole)
+
+    template = fillTextTemplate(template, {
+      user: state.settings.user.name,
+      bot: name,
+      roles: roles.reduce((prev, { role, characterId }) => {
+        prev[role] = state.novel.characters[characterId]?.name || ''
+        return prev
+      }, {} as Record<string, string>),
+    })
+
+    const totalTokens =
+      this.countTokens(template) + maxNewTokens + EMOTION_TOKEN_OFFSET
+
+    const parentEmotion =
+      selectLastLoadedCharacters(state).find(({ role }) => role === currentRole)
+        ?.emotion || ''
+
+    return {
+      template,
+      variables: {
+        emotions: emotions
+          .filter((emotion) => emotion !== parentEmotion)
+          .map((emotion) => ' ' + emotion),
+      },
+      totalTokens,
+    }
   }
 
-  public abstract completeResponse(
+  public completeResponse(
     response: NarrationResponse,
     variables: Map<string, string>,
     role: string
-  ): NarrationResponse
+  ): NarrationResponse {
+    const currentCharacterResponse = response.characters.find(
+      ({ role: characterRole }) => characterRole === role
+    )
+    const characterResponse = {
+      role: currentCharacterResponse?.role || role,
+      text: currentCharacterResponse?.text || '',
+      emotion: currentCharacterResponse?.emotion || '',
+      pose: currentCharacterResponse?.pose || '',
+    }
+    characterResponse.emotion =
+      variables.get('emotion')?.trim() || characterResponse.emotion
+    characterResponse.text = parseLLMResponse(
+      variables.get('text')?.trim() || ''
+    )
+
+    const index = response.characters.findIndex(
+      ({ role: characterRole }) => characterRole === role
+    )
+
+    return {
+      ...response,
+      characters: [
+        ...response.characters.slice(
+          0,
+          index !== -1 ? index : response.characters.length
+        ),
+        characterResponse,
+      ],
+    }
+  }
+
+  protected getDialogueHistoryPrompt(
+    state: RootState,
+    maxLines: number,
+    currentRole: string
+  ): string {
+    const messages = selectAllParentDialogues(state)
+    let prompt = ''
+    for (const message of [...messages].reverse().slice(-maxLines)) {
+      prompt += this.getDialogueLine(message, currentRole, prompt)
+    }
+    return prompt
+  }
+
+  protected getDialogueLine(
+    dialog:
+      | { type: 'response'; item: NarrationResponse }
+      | { type: 'interaction'; item: NarrationInteraction },
+    currentRole: string,
+    currentText?: string
+  ): string {
+    const temp = this.template()
+    let prevCharString = ''
+    let nextCharString = ''
+    let currentCharacterIndex
+    let characterCharacter
+    switch (dialog.type) {
+      case 'response':
+        currentCharacterIndex = dialog.item.characters.findIndex(
+          ({ role }) => role === currentRole
+        )
+        characterCharacter =
+          currentCharacterIndex !== -1
+            ? dialog.item.characters[currentCharacterIndex]
+            : {
+                text: '',
+                emotion: '',
+                pose: '',
+              }
+        if (currentCharacterIndex !== -1) {
+          prevCharString = dialog.item.characters
+            .slice(0, currentCharacterIndex)
+            .map(({ text, role }) => `{{${role}}}: ${text}`)
+            .join('\n')
+          nextCharString = dialog.item.characters
+            .slice(currentCharacterIndex + 1)
+            .map(({ text, role }) => `{{${role}}}: ${text}`)
+            .join('\n')
+        } else {
+          prevCharString = dialog.item.characters
+            .map(({ text, role }) => `{{${role}}}: ${text}`)
+            .join('\n')
+        }
+        if (dialog.item.parentInteractionId) {
+          return (
+            (prevCharString ? prevCharString + '\n' : '') +
+            (characterCharacter.text
+              ? temp.response +
+                `{{char}}'s reaction: ${characterCharacter.emotion}\n` +
+                `{{char}}: ${characterCharacter.text}\n`
+              : '') +
+            (nextCharString ? `${temp.instruction}${nextCharString}\n` : '')
+          )
+        } else {
+          return (
+            (prevCharString ? `${temp.instruction}${prevCharString}\n` : '') +
+            (characterCharacter.text
+              ? temp.response + characterCharacter.text
+              : '') +
+            '\n' +
+            (nextCharString ? `${temp.instruction}${nextCharString}\n` : '')
+          )
+        }
+      case 'interaction':
+        if (
+          (currentText?.lastIndexOf(temp.instruction) || 0) <
+          (currentText?.lastIndexOf(temp.response) || 1)
+        )
+          return `${temp.instruction}{{user}}: ${dialog.item.query}\n`
+        else return `{{user}}: ${dialog.item.query}\n`
+    }
+  }
+
+  protected getResponseAskLine(
+    state: RootState,
+    maxTokens: number,
+    role: string
+  ): string {
+    const temp = this.template()
+    const currentResponse =
+      state.narration.responses[state.narration.currentResponseId]
+    const currentCharacterResponse = currentResponse?.characters.find(
+      ({ role: characterRole }) => characterRole === role
+    )
+    const scene = selectCurrentScene(state)
+    const existingEmotion = currentCharacterResponse?.emotion || ''
+    const existingText = currentCharacterResponse?.text || ''
+    const charStops = scene?.roles
+      .map(({ role }) => {
+        return `"\\n{{${role}}}:","\\n{{${role}}}'s reaction:"`
+      })
+      .concat(temp.stops.map((stop) => `"${stop}"`))
+      .join(',')
+    return (
+      temp.askLine +
+      `{{char}}'s reaction:${
+        existingEmotion
+          ? ' ' + existingEmotion
+          : '{{SEL emotion options=emotions}}'
+      }\n` +
+      `{{char}}:${existingText}{{GEN text max_tokens=${maxTokens} stop=["\\n{{user}}:",${charStops}]}}`
+    )
+  }
 
   protected getCharacterSpecs(card: MikuCard): {
     persona: string
@@ -37,6 +247,16 @@ export abstract class AbstractPromptStrategy {
       greeting: card.data.first_mes,
       name: card.data.name,
     }
+  }
+
+  protected getRoleEmotions(state: RootState, role: string): string[] {
+    const characters = selectCurrentCharacterOutfits(state)
+    const characterEmotions =
+      EMOTION_GROUP_TEMPLATES[
+        characters.find((character) => character.role === role)?.outfit
+          ?.template || 'base-emotions'
+      ].emotionIds
+    return characterEmotions
   }
 
   protected countTokens(template: string): number {
